results: rmse outlier-KL barrier holds coherence over the loop; README + log-incoherence plot
Headline (gemma-3-4b-it s42, care-over-authority): aggregating the kl_rev
barrier by rmse over token positions (not the mean) holds coherence flat at
0.997 across all 8 rounds, where the mean aggregate collapses to 0.62 by r7
(token loops). Mean dilutes the few incoherent positions under the tau gate;
rmse is outlier-sensitive and fires on them. Cost is depth (rmse run leashes
to base, trait stays shallow); matched control still running.

- plot.py: coherence panel -> log-incoherence (1-coh, log axis, down=coherent);
  map coherence axis matches; red steer kept on the over-pipeline panels only.
- heal.py: fix kl_agg=p95 crash (torch.quantile rejects bf16 -> .float()).
- run.py: persist per-round adapter gens (adapter_gen) for the outputs table.
- config.py: coh_floor early-stop knob.
- README: results table (mean vs rmse), trajectory figure, outputs-over-loop
  appendix (per-round completions as quotes); spec persona corrected to pos-neg.
- docs/reviews: kl_agg review, pool saturation test, care-lens plan.

+++ src/steer_heal/heal.py
@@ -36,7 +36,7 @@ def _agg_kl(kl_pos, how: str):
     # +eps inside the sqrt: B=0 LoRA init makes every kl_pos exactly 0 at step 0, and bare
     # sqrt(0) has an infinite gradient (0/0), which the relu's zero-derivative turns into 0*nan.
     if how == "rmse": return (kl_pos.pow(2).mean() + 1e-8).sqrt()
-    if how == "p95": return torch.quantile(kl_pos, 0.95)
+    if how == "p95": return torch.quantile(kl_pos.float(), 0.95)  # quantile() rejects bf16
     if how == "max": return kl_pos.max()
